Remove unrelated assemblies
Each assembly is referenced by an IfcRelContainedInSpatialStructure entity. This entity is necessary for most programs to read the file. However, this entity usually contains many assemblies that we did not select which causes errors if they are not present. So, we must remove those assemblies that we did not select.