Prevent model corruption
In order to produce a valid IFC file, we were editing the references on some entities. However, this has the side effect of editing those entities in the original model. This corrupted the old model for future exports. To fix this, revert the changes after adding them to the new model.