Make widgets translateable
Create a separate ui.py file containing classes extending Qt's widget classes. These widgets are meant to behave exactly the same with the additional functionality of translating themeselves upon receiving a signal. The plan is to implement all of the widgets in this way without having to change any of the original code in the main file.